+++ lib/event_bloc.dart
@@ -1,5 +1,6 @@
 library event_bloc;
 
-export 'package:event_bloc/src/event_bloc/bloc.dart';
-export 'package:event_bloc/src/event_bloc/event_channel.dart';
-export 'package:event_bloc/src/event_bloc/repository.dart';
+export 'package:event_bloc/src/widgets/bloc_provider.dart';
+export 'package:event_bloc/src/widgets/repository_provider.dart';
+
+export 'package:event_bloc/event_bloc_no_widgets.dart';