+++ lib/event_bloc_no_widgets.dart
@@ -1,0 +1,5 @@
+library event_bloc_no_widgets;
+
+export 'package:event_bloc/src/event_bloc/bloc.dart';
+export 'package:event_bloc/src/event_bloc/event_channel.dart';
+export 'package:event_bloc/src/event_bloc/repository.dart';