+++ lib/event_bloc_widgets.dart
@@ -1,6 +1,0 @@
-library event_bloc_widgets;
-
-export 'package:event_bloc/src/widgets/bloc_provider.dart';
-export 'package:event_bloc/src/widgets/repository_provider.dart';
-
-export 'package:event_bloc/event_bloc.dart';