+++ lib/src/event_bloc/repository.dart
@@ -1,4 +1,4 @@
-import 'package:event_bloc/event_bloc.dart';
+import 'package:event_bloc/event_bloc_no_widgets.dart';
 
 class Repository implements Disposable {
   late final BlocEventChannel channel;