+++ lib/src/widgets/repository_provider.dart
@@ -1,4 +1,4 @@
-import 'package:event_bloc/event_bloc.dart';
+import 'package:event_bloc/event_bloc_no_widgets.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';