+++ test/test_classes.dart
@@ -1,0 +1,8 @@
+import 'package:event_bloc/event_bloc_no_widgets.dart';
+
+class TestBloc extends Bloc {
+  @override
+  final BlocEventChannel eventChannel;
+
+  TestBloc(this.eventChannel);
+}