+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+* Added BlocEventChannel, allowing for an event-based solution to State Management
+* Added Bloc, with integration to the BlocEventChannel
+* Added BlocProvider, which will automatically update UI that listens to it (using the provided static methods)
+* Added Repository, with integration to the BlocEventChannel
+* Added RepositoryProvider